requirements.txt
# Dependencias principales para Firebase ETL
firebase-admin>=6.0.0
python-dotenv>=1.0.0

# Dependencias para procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.13.0
shapely>=2.0.0

# Dependencias para Google Sheets y Google Drive
gspread>=5.7.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Dependencia para Google Maps API
googlemaps>=4.10.0

# Dependencias para lectura de archivos Excel
openpyxl>=3.0.0
xlrd>=2.0.0

# Dependencias para SECOP API
sodapy>=1.5.0

# Dependencias opcionales para desarrollo
requests>=2.28.0
tqdm>=4.64.0
psutil>=5.9.0
python-dateutil>=2.8.0